Split `common.ts` and moved things to `src`.
